refactor(lookalikes): split check-lookalikes.ts into cohesive modules
1598 -> 520 LOC, five new siblings, all under the 600-LOC hook threshold.
Pure extraction: code moved verbatim, only local->parameter conversion, and
every prior export is re-exported from check-lookalikes.ts so src/package.ts,
src/handlers/tools.ts, discover-brand-domains.ts and all test import sites are
untouched.

  lookalike-dns.ts            238  NS-existence filter, wildcard canary, A/MX probe
  lookalike-enrichment.ts     250  RDAP + HEAD web probe, fail-soft defaults
  lookalike-attribution.ts    240  same-entity / brand-held registration logic
  lookalike-findings.ts       382  three-axis contract + per-candidate builders
  lookalike-summary-findings  245  aggregates, scan_status notices, enumeration

assertAxisInvariants is intact: the `lookalikeDomains` key on the mail-capable
summary — the one a past change dropped while setting only `mailCapableDomains`,
which the suite caught — is carried into buildMailCapableSummaryFinding with the
reason inline, and lookalike-findings.ts restates all four axis invariants in its
module JSDoc.

Test counts identical to the pre-edit baseline across three runs (rebuilt each
time, since root test/ runs against dist/): 663 passed / 7913 passed / 13
skipped. The single failing file is test/wasm-integration.test.ts, which needs
the gitignored crates/*/pkg and fails the same way on a clean checkout.
